requirements.txt
absl-py==0.11.0
alembic==1.5.5
anyio==2.1.0
argon2-cffi==20.1.0
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrdict==2.0.1
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.2.1
bravado==11.0.2
bravado-core==5.17.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cliff==3.7.0
cmaes==0.8.2
cmd2==1.5.0
colorama==0.4.4
colorlog==4.7.2
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
entrypoints==0.3
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.12
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-resources==4.0.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
json5==0.9.5
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyter-server==1.3.0
jupyterlab==3.0.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
msgpack==1.0.2
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
neptune-client==0.9.1
neptune-contrib==0.27.1
neptune-notebooks==0.0.16
neptune-tensorflow-keras==0.9.0
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
optuna==2.5.0
packaging==20.8
pandas==1.2.1
pandocfilters==1.4.3
parso==0.7.1
path==15.1.0
path.py==12.5.0
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plotly==4.14.3
prettytable==2.0.0
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
PyJWT==2.0.1
pylint==2.6.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3987==1.3.8
rsa==4.6
scikit-learn==0.24.0
scipy==1.6.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smmap==3.0.5
sniffio==1.2.0
SQLAlchemy==1.3.23
stevedore==3.3.0
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
tb-nightly==2.5.0a20201223
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-datasets==4.1.0
tensorflow-estimator==2.4.0
tensorflow-metadata==0.26.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf-estimator-nightly==2.5.0.dev2020122301
tf-nightly-gpu==2.5.0.dev20201223
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
typing-extensions==3.7.4.3
unlzw==0.1.1
urllib3==1.26.2
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
